Separate guards & services
guards should not be considered as services since they are mostly used in routing (no need to be injected as a dependency)
